fix - app ratings controller

+++ AppStoreAPI/Models/ApplicationRatings_dbo.cs
@@ -10,7 +10,7 @@ namespace AppStoreAPI.Models
         public int id { get; set; }
         public int idUser { get; set; }
         public int idApplication { get; set; }
-        public double rating { get; set; }
+        public int rating { get; set; }
         public string comment { get; set; }
     }
 }
